Moved checkValidImageURL into utils dir
